Tweak objectives display and filtering
Previously, given key names A and B, if both A and B shared an objective, there would be a row for A, with B listed in "Others", but no row for B.

Now there is.

In addition to that, the Others column will always show the data we know about, regardless of account checkboxes' state.

Delete an unused private var.